add create_encoder_func in create_model to facilitate custom encoders; reduce code redundancy in create_encoder (patch_size, patch_channel, condition_channel, time_channel, ...), only parameters need to be pre-checked are poped;

+++ flemme/dataset/pcd.py
@@ -90,7 +90,7 @@ class PcdSegDataset(PcdDataset):
     def __getitem__(self, index):
         """Get the pcds"""
         pcd = load_pcd(self.pcd_path_list[index])
-        label = np.loadtxt(self.label_path_list[index])
+        label = np.loadtxt(self.label_path_list[index], dtype=int)
         if self.data_transform:
             n_state, t_state = get_random_state()
             pcd = self.data_transform(pcd)
